requirements.txt
# still incomplete
django-extensions
django-celery
djangorestframework
djangorestframework-filters
pika
